+++ backend/plugins/frontline_api/src/modules/inbox/graphql/schemas/conversation.ts
@@ -31,6 +31,8 @@ export const types = `
 
     messages: [ConversationMessage]
     callProAudio: String
+    callProPotentialCustomerIds: [String]
+    callProPhone: String
 
     tags: [Tag]
     customer: Customer